backend/requirements.txt
fastapi[all]
sqlalchemy
psycopg2
python-dotenv
requests
passlib[bcrypt]
python-multipart
python-jose[cryptography]

